Add files via upload
This is my first C++ project. I am creating a Database for storing information about players for Juventus, the soccer team. Not finished yet, still have some functions to work out, but this program essentially reads from a .csv file right now, and will display the info contained, as well as append into the file. I need to work on the function which will carry out editing operations, but that will come on a future screenshot.